refactor: [step1피드백] PathFInder 중복제거
