+++ src/main/java/nextstep/subway/application/PathService.java
@@ -1,7 +1,7 @@
 package nextstep.subway.application;
 
 import nextstep.subway.application.dto.PathResponse;
-import nextstep.subway.domain.PathFinder;
+import nextstep.subway.domain.pathfinder.PathFinder;
 import nextstep.subway.domain.PathType;
 import nextstep.subway.domain.Section;
 import nextstep.subway.domain.Station;
