+++ src/main/java/nextstep/subway/domain/CustomWeightedEdge.java
@@ -17,11 +17,11 @@ public class CustomWeightedEdge extends DefaultWeightedEdge {
         return duration;
     }
 
-    public void setDistance(final int distance) {
+    public void addDistance(final int distance) {
         this.distance = distance;
     }
 
-    public void setDuration(final int duration) {
+    public void addDuration(final int duration) {
         this.duration = duration;
     }
 }
