--- src/main/java/nextstep/subway/domain/PathFinder.java
@@ -1,10 +1,0 @@
-package nextstep.subway.domain;
-
-import nextstep.subway.application.dto.PathResponse;
-
-import java.util.List;
-
-public interface PathFinder {
-    boolean isType(PathType pathType);
-    PathResponse findPath(final List<Section> sections, final Station sourceStation, final Station targetStation);
-}
