+++ src/test/java/nextstep/subway/unit/PathFinderTest.java
@@ -3,6 +3,8 @@ package nextstep.subway.unit;
 import nextstep.subway.application.dto.PathResponse;
 import nextstep.subway.application.dto.StationResponse;
 import nextstep.subway.domain.*;
+import nextstep.subway.domain.pathfinder.PathFinder;
+import nextstep.subway.domain.pathfinder.ShortestDistancePathFinder;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
