+++ src/test/java/nextstep/subway/unit/PathServiceMockTest.java
@@ -7,6 +7,7 @@ import nextstep.subway.application.StationService;
 import nextstep.subway.application.dto.PathResponse;
 import nextstep.subway.application.dto.StationResponse;
 import nextstep.subway.domain.*;
+import nextstep.subway.domain.pathfinder.ShortestDistancePathFinder;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
